增加签到内容和功能性内容及其修改

+++ src/main/java/com/hqq/colorful_life/model/service/serviceImpl/CommentServiceImpl.java
@@ -23,6 +23,9 @@ public class CommentServiceImpl implements CommentService {
 
     @Override
     public int deleteByPrimaryKey(Integer id) {
+        if (id == null){
+            throw new UniteException(ExceptionEnum.REQUEST_PARAM_ERROR);
+        }
         return commentMapper.deleteByPrimaryKey(id);
     }
